Update test_combined_open_ended.py
This test fails even though the routine properly removes the inserted "onerror="eval()".  The test string (iframe_clean) has the same fields but in a different order as the input (iframe_dirty) string.  The fix is to change the test string (iframe_clean) to match the order of the input test string, minus the field that should be removed.  It now passes.